+++ R/zzz.R
@@ -1,5 +1,5 @@
 utils::globalVariables(c(
-  ".", ".N", ":=", "N", "severity", "status",
+  ".", ".N", ":=", "ID", "N", "severity", "status",
   "order", "munged_prefix", "munged_file",
   "trait_name_source", "ldsc_object_name",
   "..beta_columns", "..se_columns", "..sample_size_columns",